+++ powershell-awesome-framework/ReleaseNotes.md
@@ -4,17 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.4] - 2023.07.23
+## [0.0.4] - 2023.07.28
 
 ### Added
 
-- 
+- `Start-PAF`
 
 ### Changed
 
 - system snippets path
 - get config path data
 - `Show-PAFSnippetMenu`
+- merge read snippets to one function `Get-PAFSnippets`
 
 ## [0.0.3] - 2023.07.22
 
@@ -24,7 +25,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Changed
 
-- 
+- n/a
 
 ## [0.0.2] - 2023.07.21